refactor: Separate delete dialog and button for general use.
